refactor: extract CJK heuristics into shared LanguageHeuristics namespace
CJK detection logic was duplicated three times with subtle inconsistencies:

- AppleIntelligenceServices.swift had file-private containsCJKScalars /
  isMostlyCJKScalars / isCJKScalar helpers, plus thin private static
  wrappers (containsCJK / isMostlyCJK) on AppleIntelligenceRecipeGenerator
  that delegated to them.
- DefaultRecipeGenerator.swift had its own private static containsCJK /
  isMostlyCJK / isCJKScalar.
- WikimediaImageService.swift had its own private containsCJK with
  inline Unicode-range checks.

Same four Unicode ranges (CJK Unified Ideographs, Extension A, Japanese
kana, Hangul) in each, same 30% threshold for "mostly CJK". Now lives
in one place.

LanguageHeuristics.swift (new):
- LanguageHeuristics.containsCJK(_:) — true if any character is CJK.
- LanguageHeuristics.isMostlyCJK(_:) — true if >30% of alphabetic
  characters are CJK, used to classify text bodies.
- isCJKScalar(_:) stays private to this namespace.

All 8 call sites in the three previously-duplicated services updated
to use LanguageHeuristics.* directly. Audited via grep to confirm no
stale references remain.

What I did NOT refactor (intentionally):

- enforceLanguage near-duplicates between the generator (operates on
  InitialRecipeDraft) and the refiner (operates on RefinedRevisionDraft).
  Sharing would need a protocol abstraction across two different DTO
  shapes, which would add complexity for ~30 lines saved. Left as-is.
- LanguageModelSession.respond(to:generating:) boilerplate repeated
  ~10+ times. A generic helper would need to express the @Generable
  constraint, and getting that constraint wrong would break the build.
  The risk outweighs the 2-3 lines saved per call site. Left as-is.
