+++ ios-app/RecipeSharpener.swiftpm/DefaultRecipeGenerator.swift
@@ -43,12 +43,12 @@ struct DefaultRecipeGenerator: RecipeGenerator {
     private func inferTargetLanguage(extracted: InitialRecipeDraft, expectedDish: String?) -> String? {
         guard let desc = expectedDish?.trimmingCharacters(in: .whitespacesAndNewlines),
               !desc.isEmpty else { return nil }
-        let descIsCJK = Self.containsCJK(desc)
+        let descIsCJK = LanguageHeuristics.containsCJK(desc)
         let summary = extracted.summary
         let ingredients = extracted.ingredients.map(\.name).joined(separator: " ")
         let steps = extracted.steps.map(\.text).joined(separator: " ")
         let extractedSampleText = "\(summary) \(ingredients) \(steps)"
-        let extractedIsMostlyCJK = Self.isMostlyCJK(extractedSampleText)
+        let extractedIsMostlyCJK = LanguageHeuristics.isMostlyCJK(extractedSampleText)
 
         if descIsCJK && !extractedIsMostlyCJK {
             return "Chinese"
@@ -57,31 +57,5 @@ struct DefaultRecipeGenerator: RecipeGenerator {
             return "English"
         }
         return nil
-    }
-
-    private static func containsCJK(_ s: String) -> Bool {
-        s.unicodeScalars.contains { Self.isCJKScalar($0.value) }
-    }
-
-    private static func isMostlyCJK(_ s: String) -> Bool {
-        var cjkCount = 0
-        var letterCount = 0
-        for scalar in s.unicodeScalars {
-            if scalar.properties.isAlphabetic || isCJKScalar(scalar.value) {
-                letterCount += 1
-                if isCJKScalar(scalar.value) {
-                    cjkCount += 1
-                }
-            }
-        }
-        guard letterCount > 0 else { return false }
-        return Double(cjkCount) / Double(letterCount) > 0.3
-    }
-
-    private static func isCJKScalar(_ v: UInt32) -> Bool {
-        (0x4E00...0x9FFF).contains(v)
-            || (0x3400...0x4DBF).contains(v)
-            || (0x3040...0x30FF).contains(v)
-            || (0xAC00...0xD7AF).contains(v)
     }
 }
